- fixed compilation of WinXX agent
git-svn-id: svn://svn.zabbix.com/trunk@3888 97f52cf1-0a1b-0410-bd0e-c28be96e8082

+++ src/libs/zbxlog/log.c
@@ -256,7 +256,7 @@ void zabbix_log(int level, const char *fmt, ...)
 					}
 				}
 
-				if(old_size > buf.st_size)
+				if(old_size > (size_t)(buf.st_size))
 				{
 					redirect_std(log_filename);
 				}
